+++ pysui/sui/sui_types/event_filter.py
@@ -14,7 +14,7 @@
 """Sui query and get event filter types."""
 
 from typing import Optional, Union
-from deprecated.sphinx import versionadded
+from deprecated.sphinx import versionadded, deprecated
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_types.collections import SuiMap
 from pysui.sui.sui_types.scalars import ObjectID, SuiInteger, SuiString
@@ -24,6 +24,7 @@ class _EventFilterType:
     """Base class for identifying valid filters."""
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class SenderEventQuery(_EventFilterType, SuiMap):
     """Query events for Sender address."""
 
@@ -32,6 +33,7 @@ class SenderEventQuery(_EventFilterType, SuiMap):
         super().__init__("Sender", sender.value.value)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class TransactionEventQuery(_EventFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -40,6 +42,7 @@ class TransactionEventQuery(_EventFilterType, SuiMap):
         super().__init__("Transaction", txid)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class PackageEventQuery(_EventFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -47,6 +50,8 @@ class PackageEventQuery(_EventFilterType, SuiMap):
         """Initialize query parameter."""
         super().__init__("Package", package_id)
 
+
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class MoveEventModuleQuery(_EventFilterType, SuiMap):
     """Query events with the given move event module name."""
 
@@ -55,6 +60,8 @@ class MoveEventModuleQuery(_EventFilterType, SuiMap):
         sdict = {"module": module, "package": package_id}
         super().__init__("MoveEventModule", sdict)
 
+
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class MoveModuleEventQuery(_EventFilterType, SuiMap):
     """Query events emitted in a specified Move module."""
 
@@ -64,6 +71,7 @@ class MoveModuleEventQuery(_EventFilterType, SuiMap):
         super().__init__("MoveModule", sdict)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class MoveEventTypeQuery(_EventFilterType, SuiMap):
     """Query events for Move Event on Struct type."""
 
@@ -72,6 +80,7 @@ class MoveEventTypeQuery(_EventFilterType, SuiMap):
         super().__init__("MoveEventType", struct)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class MoveEventFieldQuery(_EventFilterType, SuiMap):
     """Query events for Move Event fields."""
 
@@ -83,6 +92,7 @@ class MoveEventFieldQuery(_EventFilterType, SuiMap):
         super().__init__("MoveEventField", field_path)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class TimeRangeEventQuery(_EventFilterType, SuiMap):
     """Query events for Time Range."""
 
@@ -92,10 +102,13 @@ class TimeRangeEventQuery(_EventFilterType, SuiMap):
         super().__init__("TimeRange", sdict)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class AndFilter(_EventFilterType, SuiMap):
     """AndFilter joins two _EventFilterTypes."""
 
-    def __init__(self, *, lhs_filter: _EventFilterType, rhs_filter: _EventFilterType) -> None:
+    def __init__(
+        self, *, lhs_filter: _EventFilterType, rhs_filter: _EventFilterType
+    ) -> None:
         """__init__ Filter initializer.
 
         Where event must be of type lhs_filter AND rhs_filter.
@@ -106,16 +119,23 @@ class AndFilter(_EventFilterType, SuiMap):
         :type rhs_filter: _EventFilterType
         :raises AttributeError: If either lhs_filter or rhs_filter not of type _EventFilterType
         """
-        if isinstance(lhs_filter, _EventFilterType) and isinstance(rhs_filter, _EventFilterType):
+        if isinstance(lhs_filter, _EventFilterType) and isinstance(
+            rhs_filter, _EventFilterType
+        ):
             super().__init__("And", [lhs_filter.filter, rhs_filter.filter])
         else:
-            raise AttributeError("AndFilter constuctor requires arguments to be of type _EventFilterType")
+            raise AttributeError(
+                "AndFilter constuctor requires arguments to be of type _EventFilterType"
+            )
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class OrFilter(_EventFilterType, SuiMap):
     """AndFilter joins two _EventFilterTypes."""
 
-    def __init__(self, *, lhs_filter: _EventFilterType, rhs_filter: _EventFilterType) -> None:
+    def __init__(
+        self, *, lhs_filter: _EventFilterType, rhs_filter: _EventFilterType
+    ) -> None:
         """__init__ Filter initializer.
 
         Where event may be of type lhs_filter OR rhs_filter.
@@ -126,12 +146,17 @@ class OrFilter(_EventFilterType, SuiMap):
         :type rhs_filter: _EventFilterType
         :raises AttributeError: If either lhs_filter or rhs_filter not of type _EventFilterType
         """
-        if isinstance(lhs_filter, _EventFilterType) and isinstance(rhs_filter, _EventFilterType):
+        if isinstance(lhs_filter, _EventFilterType) and isinstance(
+            rhs_filter, _EventFilterType
+        ):
             super().__init__("Or", [lhs_filter.filter, rhs_filter.filter])
         else:
-            raise AttributeError("OrFilter constuctor requires arguments to be of type _EventFilterType")
+            raise AttributeError(
+                "OrFilter constuctor requires arguments to be of type _EventFilterType"
+            )
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class AllFilter(_EventFilterType, SuiMap):
     """AllFilter will alllow all events in filter(s) criteria to be delivered."""
 
@@ -145,6 +170,7 @@ class AllFilter(_EventFilterType, SuiMap):
         super().__init__("All", [x.filter for x in filters])
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class AnyFilter(_EventFilterType, SuiMap):
     """AnyFilter will alllow any events in filter(s) criteria to be delivered."""
 
@@ -159,10 +185,12 @@ class AnyFilter(_EventFilterType, SuiMap):
 
 
 @versionadded(version="0.20.0", reason="Added from Sui RPC API 1.0.0")
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class _TransactionFilterType:
     """Base class for identifying valid transaction filters."""
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class CheckpointEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -171,6 +199,7 @@ class CheckpointEvent(_TransactionFilterType, SuiMap):
         super().__init__("Checkpoint", checkpoint_id)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class MoveFunctionEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -183,11 +212,18 @@ class MoveFunctionEvent(_TransactionFilterType, SuiMap):
         """Initialize query parameter."""
         package_id = package_id if isinstance(package_id, str) else package_id.value
         module_name = module_name if isinstance(module_name, str) else module_name.value
-        function_name = function_name if isinstance(function_name, str) else function_name
-        sdict = {"package": package_id, "module": module_name, "function": function_name}
+        function_name = (
+            function_name if isinstance(function_name, str) else function_name
+        )
+        sdict = {
+            "package": package_id,
+            "module": module_name,
+            "function": function_name,
+        }
         super().__init__("MoveFunction", sdict)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class InputObjectEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -197,6 +233,7 @@ class InputObjectEvent(_TransactionFilterType, SuiMap):
         super().__init__("InputObject", object_id)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class ChangedObjectEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -206,6 +243,7 @@ class ChangedObjectEvent(_TransactionFilterType, SuiMap):
         super().__init__("ChangedObject", object_id)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class FromAddressEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -215,6 +253,7 @@ class FromAddressEvent(_TransactionFilterType, SuiMap):
         super().__init__("FromAddress", address_id)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class ToAddressEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
@@ -224,21 +263,37 @@ class ToAddressEvent(_TransactionFilterType, SuiMap):
         super().__init__("ToAddress", address_id)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class FromAndToAddressEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
-    def __init__(self, from_address_id: Union[str, SuiAddress], to_address_id: Union[str, SuiAddress]):
+    def __init__(
+        self,
+        from_address_id: Union[str, SuiAddress],
+        to_address_id: Union[str, SuiAddress],
+    ):
         """Initialize query parameter."""
-        from_address_id = from_address_id if isinstance(from_address_id, str) else from_address_id.address
-        to_address_id = to_address_id if isinstance(to_address_id, str) else to_address_id.address
+        from_address_id = (
+            from_address_id
+            if isinstance(from_address_id, str)
+            else from_address_id.address
+        )
+        to_address_id = (
+            to_address_id if isinstance(to_address_id, str) else to_address_id.address
+        )
         sdict = {"from": from_address_id, "to": to_address_id}
         super().__init__("FromAndToAddress", sdict)
 
 
+@deprecated(version="0.65.0", reason="Sui dropping JSON RPC subscriptions.")
 class TransactionKindEvent(_TransactionFilterType, SuiMap):
     """Query events for Transaction."""
 
     def __init__(self, transaction_kind: Union[str, SuiString]):
         """Initialize query parameter."""
-        transaction_kind = transaction_kind if isinstance(transaction_kind, str) else transaction_kind.value
+        transaction_kind = (
+            transaction_kind
+            if isinstance(transaction_kind, str)
+            else transaction_kind.value
+        )
         super().__init__("TransactionKindEvent", transaction_kind)